refactor(tests): consolidate cascade tests into single file
Remove test_review_cascade.py which contained incorrect tests that assumed
model-level cascade (which doesn't exist by design).

Rename test_review_cascade_mixin.py -> test_review_cascade.py as the
canonical test suite for cascade functionality.

Update docstring to clarify architectural decision:
- Cascade is VIEW-LEVEL only
- Direct model method calls do NOT cascade
- Only views using CollectionReviewActionCascadeMixin trigger cascade

Result: Single, clear test file with 12 passing tests that correctly
verify the mixin-based cascade behavior.
